reorganized some tests
