+++ tests/src/session/queue.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import { Queue } from '../../src/session/queue'
+import { Queue } from '../../../src/session/queue'
 
 describe('Queue Tests', () => {
     it('should create an empty queue', () => {
